requirements.txt
apache-airflow==2.10.0
pyspark==3.5.1
pandas==2.2.2
pytest==8.3.2